+++ src/main/java/project/controller/StaticHtmlController.java
@@ -1,17 +1,18 @@
-package controller;
+package project.controller;
 
-import http.HttpRequest;
-import http.HttpResponse;
-import http.HttpStatus;
+import project.http.HttpRequest;
+import project.http.HttpResponse;
+import project.http.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 
-public class StyleSheetController implements Controller {
+public class StaticHtmlController implements Controller {
 
-    private static final Logger logger = LoggerFactory.getLogger(StyleSheetController.class);
+    private static final Logger logger = LoggerFactory.getLogger(StaticHtmlController.class);
+
     @Override
     public HttpResponse process(HttpRequest httpRequest) {
         try {
@@ -26,10 +27,9 @@ public class StyleSheetController implements Controller {
         byte[] data = Files.readAllBytes(new File("./webapp" + httpRequest.getPath()).toPath());
 
         return new HttpResponse.Builder(HttpStatus.SUCCESSFUL)
-                .header("Content-Type", "text/css")
+                .header("Content-Type", "text/html;charset=utf-8")
                 .header("Content-Length", String.valueOf(data.length))
                 .body(data)
                 .build();
     }
-
 }
