+++ src/main/java/project/controller/StyleSheetController.java
@@ -1,18 +1,17 @@
-package controller;
+package project.controller;
 
-import http.HttpRequest;
-import http.HttpResponse;
-import http.HttpStatus;
+import project.http.HttpRequest;
+import project.http.HttpResponse;
+import project.http.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 
-public class StaticHtmlController implements Controller {
+public class StyleSheetController implements Controller {
 
-    private static final Logger logger = LoggerFactory.getLogger(StaticHtmlController.class);
-
+    private static final Logger logger = LoggerFactory.getLogger(StyleSheetController.class);
     @Override
     public HttpResponse process(HttpRequest httpRequest) {
         try {
@@ -27,9 +26,10 @@ public class StaticHtmlController implements Controller {
         byte[] data = Files.readAllBytes(new File("./webapp" + httpRequest.getPath()).toPath());
 
         return new HttpResponse.Builder(HttpStatus.SUCCESSFUL)
-                .header("Content-Type", "text/html;charset=utf-8")
+                .header("Content-Type", "text/css")
                 .header("Content-Length", String.valueOf(data.length))
                 .body(data)
                 .build();
     }
+
 }
