+++ src/main/java/project/http/HttpMethod.java
@@ -1,4 +1,4 @@
-package http;
+package project.http;
 
 public enum HttpMethod {
     GET, POST, PUT, PATCH, DELETE;
