+++ src/main/java/project/http/HttpRequest.java
@@ -1,4 +1,4 @@
-package http;
+package project.http;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
