+++ src/main/java/project/http/RequestParser.java
@@ -1,8 +1,8 @@
-package http;
+package project.http;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import util.IOUtils;
+import project.util.IOUtils;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.net.URLDecoder;
