+++ src/main/java/project/util/IOUtils.java
@@ -1,4 +1,4 @@
-package util;
+package project.util;
 
 import java.io.BufferedReader;
 import java.io.IOException;
