+++ src/test/java/controller/StyleSheetControllerTest.java
@@ -1,8 +1,10 @@
 package controller;
 
-import http.HttpMethod;
-import http.HttpRequest;
-import http.HttpResponse;
+import project.controller.Controller;
+import project.controller.StyleSheetController;
+import project.http.HttpMethod;
+import project.http.HttpRequest;
+import project.http.HttpResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
